@@ -12,28 +12,28 @@
   },
   "devDependencies": {
     "@placemarkio/tokml": "^0.3.4",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/vite-plugin-svelte": "^4.0.2",
     "@tsconfig/svelte": "^5.0.4",
     "@types/topojson-client": "^3.1.5",
-    "@typescript-eslint/parser": "^8.9.0",
+    "@typescript-eslint/parser": "^8.17.0",
     "autoprefixer": "^10.4.20",
-    "eslint": "^9.12.0",
+    "eslint": "^9.16.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.45.1",
-    "globals": "^15.11.0",
-    "postcss": "^8.4.47",
+    "eslint-plugin-svelte": "^2.46.1",
+    "globals": "^15.13.0",
+    "postcss": "^8.4.49",
     "postcss-load-config": "^6.0.1",
-    "prettier-plugin-svelte": "^3.2.7",
+    "prettier-plugin-svelte": "^3.3.2",
     "simple-svelte-autocomplete": "^2.5.2",
-    "svelte": "^4.2.19",
-    "svelte-check": "^4.0.5",
-    "svelte-eslint-parser": "^0.41.1",
+    "svelte": "^5.5.0",
+    "svelte-check": "^4.1.0",
+    "svelte-eslint-parser": "^0.42.0",
     "svelte-preprocess": "^6.0.3",
-    "tailwindcss": "^3.4.14",
+    "tailwindcss": "^3.4.15",
     "topojson-client": "^3.1.0",
-    "tslib": "^2.8.0",
-    "typescript": "^5.6.3",
-    "vite": "^5.4.9"
+    "tslib": "^2.8.1",
+    "typescript": "^5.7.2",
+    "vite": "^5.4.11"
   },
   "dependencies": {
     "@mapbox/polylabel": "^1.0.2",
