@@ -35,7 +35,7 @@
       </h3>
       {#each groupLayers as { key, value }}
         <button
-          on:click={() => selectedBoundaryMap.set(key)}
+          onclick={() => selectedBoundaryMap.set(key)}
           class="flex items-center justify-center px-2 w-fit-12 h-8 text-sm font-medium rounded-md border border-gray-300 text-gray-700 dark:text-gray-300 hover:bg-gray-50 dark:hover-bg-white/20 hover:border-gray-400 dark:hover-bg-white/40 transition-colors duration-150 focus:outline-none focus:ring-2 focus:ring-blue-400 focus:ring-offset-2"
         >
           <span>{value.name}</span>
