@@ -1,7 +1,7 @@
 <div
   class="bg-gray-200 dark:bg-neutral-800 w-full h-1 inline-block relative overflow-hidden rounded-sm"
 >
-  <div class="loader bg-blue-500 w-48 h-1 absolute top-0 left-0" />
+  <div class="loader bg-blue-500 w-48 h-1 absolute top-0 left-0"></div>
 </div>
 
 <style>
