@@ -1,4 +1,6 @@
 <script lang="ts">
+  import { run } from 'svelte/legacy';
+
   import SidebarHeader from './SidebarHeader.svelte';
   import {
     mapStore,
@@ -15,8 +17,8 @@
   import { resetZoom } from '../../helpers/helpers';
   import PolygonLookup from 'polygon-lookup';
 
-  let districtsIntersectingAddress: Feature[];
-  let isLoading = false;
+  let districtsIntersectingAddress: Feature[] = $state();
+  let isLoading = $state(false);
   let lookup: Object | null = null;
 
   function queryAllDistrictsForCoordinates(lngLat: LngLat) {
@@ -44,32 +46,36 @@
     resetZoom($mapStore);
   }
 
-  $: if ($mapStore && $selectedCoordinates && $boundaries) {
-    queryAllDistrictsForCoordinates($selectedCoordinates);
-    $mapStore.flyTo({ center: $selectedCoordinates, zoom: 13 });
+  run(() => {
+    if ($mapStore && $selectedCoordinates && $boundaries) {
+      queryAllDistrictsForCoordinates($selectedCoordinates);
+      $mapStore.flyTo({ center: $selectedCoordinates, zoom: 13 });
 
-    if ($addressMarker) $addressMarker.remove();
-    if ($coordinatesMarker) $coordinatesMarker.remove();
-    $coordinatesMarker = new maplibregl.Marker({ color: '#2463eb' })
-      .setLngLat($selectedCoordinates)
-      .addTo($mapStore);
-  }
-
-  $: if ($mapStore) {
-    if ($isSelectingCoordinates) {
-      $mapStore.getCanvas().style.cursor = 'crosshair';
-
-      $mapStore.once('click', e => {
-        selectedCoordinates.set({
-          lat: e.lngLat.lat.toFixed(5),
-          lng: e.lngLat.lng.toFixed(5)
-        });
-        isSelectingCoordinates.set(false);
-      });
-    } else {
-      $mapStore.getCanvas().style.cursor = '';
+      if ($addressMarker) $addressMarker.remove();
+      if ($coordinatesMarker) $coordinatesMarker.remove();
+      $coordinatesMarker = new maplibregl.Marker({ color: '#2463eb' })
+        .setLngLat($selectedCoordinates)
+        .addTo($mapStore);
     }
-  }
+  });
+
+  run(() => {
+    if ($mapStore) {
+      if ($isSelectingCoordinates) {
+        $mapStore.getCanvas().style.cursor = 'crosshair';
+
+        $mapStore.once('click', e => {
+          selectedCoordinates.set({
+            lat: e.lngLat.lat.toFixed(5),
+            lng: e.lngLat.lng.toFixed(5)
+          });
+          isSelectingCoordinates.set(false);
+        });
+      } else {
+        $mapStore.getCanvas().style.cursor = '';
+      }
+    }
+  });
 </script>
 
 <SidebarHeader
