@@ -1,7 +1,8 @@
 import './app.css';
 import App from './App.svelte';
+import { mount } from 'svelte';
 
-const app = new App({
+const app = mount(App, {
   target: document.getElementById('app')
 });
 
